api/requirements_with_local_AI.txt
fastapi
uvicorn[standard]
sentence-transformers
scikit-learn
numpy
python-dotenv
requests